+++ tests/conftest.py
@@ -9,6 +9,8 @@ from tempfile import mkstemp, mkdtemp
 @pytest.fixture(scope="function")
 def tmp(request):
     class Tmp(object):
+        cleanup = True
+
         def __init__(self):
             self.dir = mkdtemp()
             _, self.file = mkstemp(dir=self.dir)
@@ -16,7 +18,7 @@ def tmp(request):
     yield t
 
     # cleanup
-    if os.path.exists(t.dir):
+    if os.path.exists(t.dir) and t.cleanup is True:
         shutil.rmtree(t.dir)
